# docs/openapi/webserver.swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "WebServer",
    "version": "1.0"
  },
  "paths": {
    "/api/health": {
      "get": {
        "tags": [
          "WebServer"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/processes": {
      "get": {
        "tags": [
          "WebServer"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/works": {
      "get": {
        "tags": [
          "WebServer"
        ],
        "parameters": [
          {
            "name": "workId",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "model",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "serial",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "process",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/works/{workId}": {
      "get": {
        "tags": [
          "WebServer"
        ],
        "parameters": [
          {
            "name": "workId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/segments/{segmentId}/download": {
      "get": {
        "tags": [
          "WebServer"
        ],
        "parameters": [
          {
            "name": "segmentId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/segments/{segmentId}/stream": {
      "get": {
        "tags": [
          "WebServer"
        ],
        "parameters": [
          {
            "name": "segmentId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/tusd/hooks/completed": {
      "post": {
        "tags": [
          "WebServer"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadCompletedRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/jobs/cleanup": {
      "post": {
        "tags": [
          "WebServer"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/jobs/archive": {
      "post": {
        "tags": [
          "WebServer"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/files/": {
      "post": {
        "tags": [
          "tusd"
        ],
        "summary": "tusd: Create an upload",
        "description": "Resumable upload endpoint (tus protocol). This route is served by tusd via reverse-proxy, not by the ASP.NET webserver.\n\nReverse proxy constraints (compose/nginx.conf):\n- Requires X-Api-Key matching TUSD_UPLOAD_API_KEY\n- Uploads allowed only from private networks/localhost\n\nClient flow: POST /files/ -> Location: /files/{id} -> PATCH /files/{id} until Upload-Offset == Upload-Length.",
        "security": [
          {
            "TusdUploadApiKey": []
          }
        ],
        "parameters": [
          {
            "name": "Tus-Resumable",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1.0.0"
            }
          },
          {
            "name": "Upload-Length",
            "in": "header",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "Upload-Metadata",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created. Location header contains the upload URL.",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                },
                "description": "Upload URL (typically /files/{id})"
              }
            }
          },
          "401": {
            "description": "Missing/invalid X-Api-Key (reverse proxy)"
          },
          "403": {
            "description": "Upload not allowed from this network (reverse proxy)"
          }
        }
      }
    },
    "/files/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "head": {
        "tags": [
          "tusd"
        ],
        "summary": "tusd: Get upload offset",
        "security": [
          {
            "TusdUploadApiKey": []
          }
        ],
        "parameters": [
          {
            "name": "Tus-Resumable",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1.0.0"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns current Upload-Offset",
            "headers": {
              "Upload-Offset": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          },
          "404": {
            "description": "Upload not found"
          },
          "401": {
            "description": "Missing/invalid X-Api-Key (reverse proxy)"
          },
          "403": {
            "description": "Upload not allowed from this network (reverse proxy)"
          }
        }
      },
      "patch": {
        "tags": [
          "tusd"
        ],
        "summary": "tusd: Upload chunk",
        "description": "Send a chunk using tus protocol. Content-Type must be application/offset+octet-stream.",
        "security": [
          {
            "TusdUploadApiKey": []
          }
        ],
        "parameters": [
          {
            "name": "Tus-Resumable",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1.0.0"
            }
          },
          {
            "name": "Upload-Offset",
            "in": "header",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "example": "application/offset+octet-stream"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/offset+octet-stream": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "No Content. Upload-Offset header indicates new offset.",
            "headers": {
              "Upload-Offset": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          },
          "409": {
            "description": "Offset conflict"
          },
          "404": {
            "description": "Upload not found"
          },
          "401": {
            "description": "Missing/invalid X-Api-Key (reverse proxy)"
          },
          "403": {
            "description": "Upload not allowed from this network (reverse proxy)"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "TusdHookRequest": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "upload": {
            "$ref": "#/components/schemas/TusdHookUpload"
          }
        },
        "additionalProperties": false
      },
      "TusdHookStorage": {
        "type": "object",
        "properties": {
          "path": {
            "type": "string",
            "nullable": true
          },
          "RelativePath": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TusdHookUpload": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "storage": {
            "$ref": "#/components/schemas/TusdHookStorage"
          },
          "metaData": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "nullable": true
          },
          "size": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UploadCompletedRequest": {
        "type": "object",
        "properties": {
          "segmentUuid": {
            "type": "string",
            "nullable": true
          },
          "workId": {
            "type": "string",
            "nullable": true
          },
          "model": {
            "type": "string",
            "nullable": true
          },
          "serial": {
            "type": "string",
            "nullable": true
          },
          "process": {
            "type": "string",
            "nullable": true
          },
          "segmentIndex": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "recordedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "durationSec": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "sizeBytes": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "localFilePath": {
            "type": "string",
            "nullable": true
          },
          "sha256": {
            "type": "string",
            "nullable": true
          },
          "tusd": {
            "$ref": "#/components/schemas/TusdHookRequest"
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "TusdUploadApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key",
        "description": "Upload API key enforced by reverse-proxy for /files/* (TUSD_UPLOAD_API_KEY)"
      }
    }
  }
}